Extract BatchCollector into its own file.
Move batch collection, flush, and retry logic out of controller.go to keep the controller focused on orchestration.
